Move LLMCostCard and LLMConsumptionCalculatorBase into llm_cost.py
