+++ council/llm/__init__.py
@@ -8,7 +8,8 @@ from .llm_config_object import LLMProvider, LLMConfigObject, LLMConfigSpec, LLMP
 from .llm_answer import llm_property, LLMAnswer, LLMProperty, LLMParsingException
 from .llm_exception import LLMException, LLMCallException, LLMCallTimeoutException, LLMTokenLimitException
 from .llm_message import LLMMessageRole, LLMMessage, LLMMessageTokenCounterBase
-from .llm_base import LLMBase, LLMResult, LLMConfigurationBase, LLMConsumptionCalculatorBase, LLMCostCard
+from .llm_base import LLMBase, LLMResult, LLMConfigurationBase
+from .llm_cost import LLMCostCard, LLMConsumptionCalculatorBase
 from .llm_fallback import LLMFallback
 from .llm_middleware import (
     LLMRequest,
