+++ council/llm/openai_chat_completions_llm.py
@@ -4,12 +4,18 @@ from typing import Any, Dict, List, Mapping, Optional, Protocol, Sequence
 
 import httpx
 from council.contexts import Consumption, LLMContext
+from council.llm import (
+    ChatGPTConfigurationBase,
+    LLMBase,
+    LLMCallException,
+    LLMConsumptionCalculatorBase,
+    LLMCostCard,
+    LLMMessage,
+    LLMMessageTokenCounterBase,
+    LLMResult,
+)
 
 from ..utils import truncate_dict_values_to_str
-from . import ChatGPTConfigurationBase
-from .llm_base import LLMBase, LLMConsumptionCalculatorBase, LLMCostCard, LLMResult
-from .llm_exception import LLMCallException
-from .llm_message import LLMMessage, LLMMessageTokenCounterBase
 
 
 class Provider(Protocol):
